feat: create Advanced Settings
